Move the documentation builder to the android slave.
While the documentation is being built, all other slaves can start building
the next PR. Since the android build is currently much shorter than the linux
builds, it can start later (once the documentation build is finished) and
still finish before the other builds. This reduces the end-to-end time by
two or three minutes if there is a queue of pending PRs.