Skip only Android states, not nodes, on Travis
Instead of skipping all node types that include Android, use the pillar
to skip Android states on Travis.

Providing 'travis: True' in the pillar for a node will enable skipping
Android states; the key can simply be omitted if not on Travis, and Salt
will fall back to running Android states by default.