Fix usage of homebrew taps
When installing autoconf213 from the homebrew/versions tap, salt
correct installs the package but fails to verify its installation
because it compares 'homebrew/versions/autoconf213', which is the
specified package name, with 'autoconf213', which is the installed
package name. Use the brew.install module function as a workaround
until this functionality is added to pkg.installed or pkgrepo.

See: https://github.com/saltstack/salt/issues/26414