Add Vagrantfile for local debugging
Add a Vagrantfile that is mostly auto-generated from .travis.yml.
Meant to be used for local debugging.
Creates a multiple-machine setup with one machine for each supported
node type, and runs the highstate as the provision step.

Note: currently doesn't work due to bugs in the vagrant salt
provisioner, which I have put PRs in to fix.