Use Salt to manage increased maxfiles on OS X
These settings are necessary to handle running the WPT (web platform
tests); taken out of the wiki to reduce the need for manual
interventions.

Note, /etc/launchd.conf is not included because OS X no longer uses
this file and it has no effect. Whether a replacement for this
functionality is required is TBD.