Pin virtualenv pip package to pin installed binary
A recent update of the virtualenv package changed the installed
binaries from `virtualenv` and `virtualenv-3.4`, to `virtualenv` and
`virtualenv-3.5`, breaking the homu installation which needs a python3
version of virtualenv. Pin the package to the newest version and update
the homu installation to use the virtualenv-3.5 binary.

Note: the python3 apt package still installs python3.4, but the version
difference should be ok since virtualenv is just used to install homu,
not run it.