Add check for Python 3.5+ in test script
The test script uses features introduced in Python 3.5 (namely
os.path.commonpath), but the shebang only specifies python3. Because
Python3.5 is fairly new, it's likely that python3 is not yet Python 3.5
(or later) for many people, so explicitly check for this.

This occured on Travis, requiring an explicit 'python: 3.5' config,
which is what spurred this commit.