Add test to check Buildbot master config
This test can be run via `./test.py sls.buildbot.master`. It will only
work properly once the Buildbot master states have been run, and is
meant to be run either in Vagrant or on Travis, not locally.

Running `./test.py` without arguments will only run the linting tests, since
those can be run anywhere with just the source code.

For now, hard-code this test to be run on the servo-master1 node on Travis.
This should be replaced with a more intelligent test dispatcher in `test.py`
which uses the top file via Salt's state.show_top in the future, but hard
coding this for now will reduce Buildbot's fragility.