Add test to check if Nginx is serving requests
urllib's error handling is not very helpful for this test, but I'm
using it in an effort to avoid extra dependencies. This only checks
if Nginx is alive and returning requests, not whether the HTTP status
codes indicate success. The reason for this is that the test pillars
we have in the repo are not valid, causing Buildbot and Homu to die
after starting, and thus Nginx returns 502 Bad Gateway errors when it
tries to reverse proxy requests.

Enable the test on Travis as well.