Create separate file for steps
In order to move the build steps to the main servo repository, they must
be converted from python code that requires the Buildbot library to
be installed (to manually build steps) into more declarative steps. This
commit starts this conversion by using heuristics to convert a list of
shell commands (akin to those in a .travis.yml or .appveyor.yml file)
into the appropriate Buildbot Steps.

The doc factory and windows factory still use hardcoded steps;
see the comments in the code for more details.

The android nightly factory now uses a relative instead of absolute path
for the .apk file to make the Jinja templating unnecessary.