Use a YAML file to drive dynamic builders
As a stepping-stone to having builder steps live in the main Servo repo,
switch to a simple data format to express builder configuration (lists
of strings accessed via yaml.safe_load) instead of needing to run or
import arbitrary untrusted Python code.

Currently, the YAML file is still part of the master configuration, but
it is now ready to be moved to the main Servo repo.