Remove duplicate testing steps for speed
Many of the tests that are just lints are being run multiple times on
different builders, but they only need to be run once. Run them on
linux-dev as it's the fastest builder.

This is also meant as an example of how much easier a YAML file makes
it to edit our build scripts; note the lack of needing to know about
`logfiles` or other parameters which have been forgotten in the past.