Tweak testing steps
Use linux-dev to run all lints because it is the fastest builder; these
are platform-independent.

Ensure that the lockfile_changed.sh and manifest_changed.sh scripts are
run for every builder, since these are platform-dependent. Note: these
should also be added for the Windows builder.

This is also meant as an example of how much easier a YAML file makes
it to edit our build scripts; note the lack of needing to know about
`logfiles` or other parameters which have been forgotten in the past.