Re-add pkill step
This automatically runs pkill for each build. It is cleaner to put this
in the main configuration instead of the lists of steps because this is
a CI-related cleanup task, not a necessary step to run the build + test
suite. Also, this ensures that the pkill step cannot be forgotten when
adding new builders.

Note: this doesn't work on Windows yet.