Ensure Buildbot master files are owned by servo
If the Buildbot master is ever erroneously started as root, the
permissions on a variety of files in the Buildbot master dir are
clobbered and the Buildbot master process is unable to access them.
Use a file.directory state to set ownership to servo:servo for the
whole folder, so that a highstate will restore the correct permissions
in case this happens.