Use explicit masterless args, not minion config
Previously, the minion config was used to specify that a local set
of file trees should be used. However, in order to test that the Salt
master states function, and to more closely match the actual deployment
of our minions, amend the minion config to point to a localhost master.
Additionally, amend our Travis config and Vagrantfile to explicitly
pass options to set the local file and pillar trees, and stop syncing
these via Vagrant as they are no longer necessary.

In the future, this will allow adding a smoke test that check if the
master can test.ping a local minion (on the same VM). However, this
requires the minion to be started, which requires that the Salt minion
also be salted, (which requires this PR to backport the macservice.py
execution module).