Use explicit masterless args, not minion config
Previously, the minion config was used to specify that a local set
of file trees should be used. However, in order to test that the Salt
master states function, and to more closely match the actual deployment
of our minions, amend the minion config to point to a localhost master.
Additionally, amend our Travis config and Vagrantfile to explicitly
pass options to set the local file and pillar trees.
Only sync them via Vagrant shared folders for the Salt master, which
should have them available to allow debugging. Because using Salt is
the only way to learn which minions (ids) would have a Salt master
running on them, hardcode the salt-master\d+ regex into the Vagrantfile.

In the future, this will allow adding a smoke test that check if the
master can test.ping a local minion (on the same VM). However, this
requires the minion to be started, which requires that the Salt minion
also be salted, (which requires this PR to backport the launchctl.py
execution module).