Enable gitfs for Salt master
Use gitfs backed by the saltfs repo (https://github.com/servo/saltfs/)
as the main source for the Salt file tree.

For local testing, additionally configure a local directory in /tmp
as a Salt file tree root, and give it precedence over the gitfs tree.
This allows temporarily making changes in the local directory to iterate
on them. Add an ADMIN_README file with some ground rules about proper
usage of the directory; this file should be set as immutable, but Salt
does not yet have a module/state for this.