Disable Homebrew Analytics
This is not the most robust of solutions, but will do the job.
A better solution would be a custom Salt state, backed by a custom
Salt module, that uses the `brew analytics` command to stay decoupled
from implementation details.

Instead, directly change the git config settings of the local
Homebrew repo to disable analytics.