Ensure specific Android SDK platform version is installed
Update the creates parameter to the cmd.run state which installs the
Android SDK to check for both the platform tools and the specific
version of the platform we are using (currently, android-18). This
makes sure that the SDK will reinstalled on future Salt runs if either
directory is missing, instead of just checking for the platform tools.

Also, extract the Android platform into a Jinja variable.

Note that this feature of Salt is not documented, hence why I didn't
do this originally - I thought it didn't exist. I only discovered it
when I went into the Salt codebase to implement it myself.