Use reduced set of builders for try builds
Running all the builds for try builds is unnecessary and uses up extra
resources, slowing down the rate of queue drainage. In particular,
the recently-added linux-rel-gczeal build takes a long time, but is not
yet gated on, and adds no value being run for try builds. Reduce the set
of builders for try builds down to a minimal set that still has full
platform coverage.