Test packaging on android builder
Our android-nightly builder has been busted for a long time, which meant
our packaging code was not being exercised. Also, the nightly builder
only runs once a day instead of on each build, which recently allowed
a regression to slip in due to an uncaught syntax error. To prevent
future regressions, test packaging on the android builder, which runs
on try and auto builds.