Add Buildbot ←→ GitHub status integration
Buildbot 0.8.12 has a GitHub status integration; it would be better to
use that directly instead of making Homu act as a middleman. Note that
this commit does not change the Homu configuration, but only adds a
GitHub status integration for Buildbot.

Note that this requires a new secret value in the Pillar - a token
that Buildbot can use to access the GitHub status API.

See http://docs.buildbot.net/current/manual/cfg-statustargets.html#githubstatus
for more details on the integration.