Test rustc download code on arm32 builder
We recently had a bug that caused our rustc download code to be broken,
but this was not caught because the builders simply used rustc from
cache instead of downloading fresh. Futhermore, the breakage was only
for cross target downloads, not regular downloads. Hence, unset
SERVO_CACHE_DIR on the arm32 builder in order to force downloading the
host and cross stdlibs for each build (cross builders download both).
(arm32 is the fastest cross builder at ~30 minutes.)