Use native Buildbot logging
We used to use a regular `print` statement for logging here, but that
output was not going anywhere we could find. Buildbot has a native
buildbot.config.error() function that we could use, but that would mean
that Buildbot would completely fail to start in case of a
misconfiguration, blocking other jobs from being run on builders that
are still correctly configured (e.g. when testing out a new builder).
Instead, continue to handle and log step misconfiguration errors at
runtime, but use native Buildbot logs to do so, generating a log file
that is accessible in the Buildbot web UI.