Return try/except to DynamicServoFactory
This was recently removed due to other bugs that were causing our error
recovery reporting to be useless, but now that those are fixed this can
be returned to allow Buildbot to start even if certain builders have
bad configurations (e.g. a new builder that is getting up to speed),
allowing other changes to land in the meantime.