Add fake Github Homebrew Token for Travis
We recently added a new homebrew nightly builder, which pushes to the
`servo/homebrew-servo` repository to make new binary nightlies
available via Homebrew. This requires a Github token to push to the
repository.

However, the original PR did not include a fake token for the Travis
failure, causing the builder to fail only on Travis (where the test
pillars did not have the corresponding token), but not in production.