Add homebrew_analytics module for robust disabling
Originally, a raw call to the `git.config` state was used to disable
Homebrew Analytics. However, this is fragile and recently broke with the
Homebrew 1.0.0 release, which migrated the Homebrew repository location.
Now that we have support for custom modules, add a `homebrew` execution
module to allow running arbitrary Homebrew commands, and a
`homebrew_analytics` module on top which allows robustly disabling
Homebrew Analytics by using the builtin `brew analytics` command.

The execution module is named `homebrew` because the built-in module
[`mac_brew.py`] is not available (hidden behind the `pkg`
__virtualname__), so its methods cannot be called.