Be robust to upstream changes when building old revs
When running builds that are not from scratch, it is possible that
upstream changes have caused old revisions on master to stop building.
Because homu maintains the invariant that each merge to master is
passing, and we are only building the old revision to create a baseline
to test the new revision, it's safe to ignore failures in the old rev.