Invalidate Salt cache in-between runs
For some reason it seems Salt is not invalidating the `_modules` cache,
causing build failures on upgrade builds (i.e. the second build on a
not-from-scratch run). Invalidate the cache manually.

Note that `nullglob` is set for safety, but we always expect `rm` to
find cache files to invalidate. In case it does not find any files, the
glob will expand to nothing, which is considered an error, causing
failure via `errexit`. That is, not ignoring the result of rm is
intentional. (`nullglob` is nicer than `failglob` because it is
friendlier to loops over globs.)