Bump Salt to 2016.3.3
Homebrew recently changed their packaging of Salt and Python, using a
virtualenvs for Salt. Because we were pinning an old version of Salt,
but not the rest of Homebrew, this caused problems where pip was not
being resolved properly. Update Salt to the latest version to fix this.

As part of the update, also update some states to avoid deprecated
parameters in favor of newer ones:
- archive.extracted: archive_user -> user, group
- cmd.run: user, group -> runas

Note that due to https://github.com/saltstack/salt/pull/36552,
`archive.extracted` states still have the archive_user argument to
ensure parent directories are made with the correct permissions.