Bump Salt to 2016.3.3
Homebrew recently changed their packaging of Salt and Python, using a
virtualenvs for Salt. Because we were pinning an old version of Salt,
but not the rest of Homebrew, this caused problems where pip was not
being resolved properly. Update Salt to the latest version to fix this.

As part of the update, also update some states to avoid deprecated
parameters in favor of newer ones:
- archive.extracted: archive_user -> user, group
- cmd.run: user, group -> runas

Note that due to https://github.com/saltstack/salt/pull/36552,
`archive.extracted` states still have the archive_user argument to
ensure parent directories are made with the correct permissions.

saltfs-migration: Upgrade Salt on all machines. Steps:
- Stop all salt-master and salt-minion services. Because we're upgrading
  Salt by two major versions at once, there are no guarantees that the
  masters and minions with different versions will be able to
  communicate.
- Use the install script to re-install Salt on each machine, the master
  first then the minion.
- Restart the salt-master and salt-minion services; pay attention to the
  system logs in case of startup failure.
- Do NOT run a highstate until all minions and masters have been updated
  to Salt 2016.3.0.