Don't link Homebrew OpenSSL, use env vars
OpenSSL is a "keg-only crate", which Homebrew 1.0.0 won't link because
```
we may end up linking against the insecure, deprecated system OpenSSL
while using the headers from Homebrew's openssl.
```

Instead, use the `OPENSSL_INCLUDE_DIR` and `OPENSSL_LIB_DIR` to
explicitly pass the paths to the Homebrew OpenSSL. Note that these paths
are currently hardcoded because they are unlikely to change.

For some reason, Homebrew was reporting this error on its stderr
but Salt was not picking it up as and failing the state - possibly
because the exit status seems to have been 0.

saltfs-migration: Run `brew unlink --dry-run openssl` on the Mac
builders, then `brew unlink openssl` once confirming the output is as
expected. This will unlink openssl on the existing builders.