Unlink existing Salt in Homebrew before installing
If a different version of Salt is already installed via Homebrew, it
will not allow installing a different version without previously
unlinking the old one. Ideally there would be an atomic operation to do
both of these in one command, but it appears Homebrew does not have one.
Thus, users of this script should be careful to watch the output in case
the script is interrupted and no Salt installation is left linked.