Fix autoconf/autoconf213 installation via Homebrew
Use an idempotent cmd.run to tap `homebrew/versions` only if necessary.

Homebrew 1.0.0 apparently tightened its return codes and will exit with
a non-zero status if linking fails during installation; there is no way
to install without linking or install and link --overwrite in one
operation. This makes it difficult to install multiple versions that
have conflicting links from the CLI, so add a custom script to handle
ensuring that autoconf and autoconf213 are installed and linked
properly, with autoconf's links over autoconf213's links,
in an idempotent way and doing as little as possible.