Temporarily disable servo-mac1 builder
The upgrade to Salt 2016.3 broke Salt highstate on the macs, so I've
been using servo-mac1 to investigate as it was already on the fritz.
However, it seems that it started returning results to Buildbot, failing
all builds, so it was turned off. Temporarily remove it from the builder
list to ensure it doesn't interfere with the (rapidly growing) queue
while I use it to investigate.