Treat './mach package' as a steps.Compile
This command is used during the Android APK build, but even though it
failed, Buildbot still went on to run the next step, because it used to
be run as a simple ShellCommand (as per issue #372).

Instead, we now check for it on the `make_step` function and use
`steps.Compile` to create the step as it is not only a closer semantic
match (we're "compiling" the APK) but it also makes it so when it fails
Builbot doesn't ignore it or move on to the next step.