Treat './mach package' as a steps.Compile
Even when this command fails, Buildbot still goes on to run the next
step, because it is run as a simple ShellCommand.

Instead, we now check for it in the `make_step` function and use
`steps.Compile` to create the step as it is not only a closer semantic
match (we're "compiling" the package) but it also makes it so when it
fails Builbot doesn't ignore it or move on to the next step.