Add step to ensure Buildbot checks out right commit
We are using `retryFetch=True` to have Buildbot retry git fetches if
they error out, as we often encounter intermittent errors (e.g. network
problems).

However, Buildbot will allow not only the first fetch + checkout to
fail, but also the second (final) fetch + checkout to fail as well
(which I consider a bug in Buildbot), making it possible for Buildbot
to run a build on the wrong revision, causing confusion.

Retrying failed fetches is too useful to disable.
Hence, to work around this, add a custom step to all builds that checks
if the commit we requested (the `revision` property) matches the
actual commit Buildbot checks out, and fail the build if not the case.