Re-add logic to provide Homebrew upload token
This was added to the previous `make_step` implementation,
but apparently was not copied over to this implementation,
causing our Homebrew uploads to fail for the last 4 days.
Re-add this logic.

Note that because adding the token to the environment and telling
Buildbot not to log the environment are configured together,
the token was not leaked during this time.