Move most Salt docs in tree
This also includes Vagrant information,
and should cover all the steps of a normal Salt workflow.

This does not (yet) include information about
setting up new Salt masters/minions,
or how to restart various services cleanly after a deploy.

Moving docs in-tree ensures they stay up to date with code changes,
as opposed to having to remember to update the wiki after long
review cycles.
The corresponding sections will be removed from the wiki
and replaced with pointers to the in-tree docs.