Disable Python bytecode caching for Buildbot master
For some reason (likely due to our "graceful" restarts of Buildbot),
when Buildbot is restarted after its configuration is updated,
it does not read the updated `.py` files but instead uses the existing
compiled bytecode (`.pyc`) files, which reflect an older version and are
not up to date. These bytecode files are also not regenerated,
meaning the new configuration does not (fully) take effect.

To work around this, disable bytecode caching for the Buildbot master
entirely to avoid using out-of-date bytecode.