Factor out python states, include sls as needed
Move the Python (2 and 3) related states into a separate SLS file,
so they can be reused without pulling in the servo user state,
which is in common.

Instead of adding this new SLS file to the top file,
include it from other SLS files that depend on a Python installation.

This makes it possible to `state.apply servo-build-dependencies`
independently, with is needed for `mach bootstrap`.

Note: No effort was made to separate Python 2 and Python 3,
as the virtualenv pip package will create executables for both.