Remove obsolete Homebrew code
This information about the Homebrew user was needed for a workaround
which involved invoking Homebrew directly,
since Homebrew cannot be run as root.

However, the improved Homebrew versions support means this workaround
(to installing autoconf and autoconf@2.13 concurrently)
is no longer necessary and was removed in
commit a0632484043aba1cbc1a0bb6ce45420f5a090d7d.

This code was inadvertently left dangling, and is now cleaned up.