Only reference sources.list.d on fully managed nodes
The reference to /etc/apt/sources.list.d allows cleaning that directory
of extraneous sources files. This should only be done on fully managed
nodes, so guard this requisite with a pillar check.

The check defaults to True for backwards compatibility
(i.e. not having to change the production pillars).