Concurrently install Android r25.2.3 tools
Install a new version of the Android tools along with the old one.
Keep the old version as the current version for now to avoid breaking
builds, but still make the new version available.
This will also make future upgrades more smooth as well.

Switch to the new URL scheme for SDK downloads,
which also has switched to using zip files instead of .tar.gz files.
Note that the new packaging format
no longer has a top-level `android-sdk-linux` folder.

Move the `build_tools` and `platform` attributes under `sdk`,
as they are only used with the SDK-related states.

Reinstate `if_missing` for the SDK `archive.extracted` state to ensure
that the new archive will be extracted if the old one is present,
so that the SDK has the correct directory layout,
This also means ownership will only be updated for that file
(the `android` binary), so add a `file.directory` state to handle
fixing the ownership for the rest of the extracted files,
and fixup the requisites as necessary.