Update expect script to propagate return codes
Unfortunately, the `android` tool always returns a 0 exit code even in
case of failure, so this does not yet catch failures.
However, the new `sdkmanager` tool seems to return appropriate exit
codes, so this will kick in in the future.

This will prevent errors from passing silently,
such as bad executable bits or incorrect file permissions.