Set a maximum priority for builds.
The whole thing about having non-infinite priorities is to facilitate
sherriffing as discussed in https://github.com/servo/homu/issues/111

These changes set a max_priority for Homu and bump Homu to the version which
respects it.

I pulled 9001 out of thin air because I consider it easy to memorize.