Allow specifying env vars in buildbot_steps.yml
Support a top-level `env` key for environment variables that are
global (should apply to all builders).
Additionally, support providing separate `env` and `commands` keys
for a builder, instead of just a list of commands.

Values for `env` should be dictionaries of string vars
representing the environment variables and their values.