Install full system Python 3 for testing
Travis has changed their pyenv configuration such that the default
Python comes first in the PATH, instead of a custom Python 3.5.
It's a good idea to use the system Python for a few reasons:
  - pyenv is not used on macOS, and we don't have it inside Docker
  - This makes it easier to run on stock hosts without pyenv,
    which is useful for running tests on e.g. our real build machines

Unfortunately, the Ubuntu images do not have the necessary python3
virtualenv package installed, and the name differs between Ubuntu
versions, so these configs require some additional snowflake steps.

Use the Salt python SLS file to ensure that the system Python is
configured appropriately before running the test suite,
which allows encapsulating the complexity into the Salt code for reuse.
Make sure to install Salt first if it is not already installed.